+++ app/src/main/kotlin/com/designdrivendevelopment/kotelok/application/AppComponent.kt
@@ -1,15 +1,15 @@
 package com.designdrivendevelopment.kotelok.application
 
 import android.content.Context
-import com.designdrivendevelopment.kotelok.CardsLearnableDefinitionsRepository
-import com.designdrivendevelopment.kotelok.DictWordDefinitionRepositoryImpl
-import com.designdrivendevelopment.kotelok.DictionariesRepositoryImpl
-import com.designdrivendevelopment.kotelok.EditWordDefRepositoryImpl
-import com.designdrivendevelopment.kotelok.PairsLearnableDefinitionsRepository
 import com.designdrivendevelopment.kotelok.R
-import com.designdrivendevelopment.kotelok.StatisticsRepositoryImpl
-import com.designdrivendevelopment.kotelok.WriterLearnableDefinitionsRepository
 import com.designdrivendevelopment.kotelok.persistence.database.KotelokDatabase
+import com.designdrivendevelopment.kotelok.repositoryImplementations.dictionariesRepository.DictionariesRepositoryImpl
+import com.designdrivendevelopment.kotelok.repositoryImplementations.dictionaryWordDefinitionsRepository.DictWordDefinitionRepositoryImpl
+import com.designdrivendevelopment.kotelok.repositoryImplementations.editWordDefnititionsRepository.EditWordDefRepositoryImpl
+import com.designdrivendevelopment.kotelok.repositoryImplementations.learnableDefinitionsRepository.CardsLearnableDefinitionsRepository
+import com.designdrivendevelopment.kotelok.repositoryImplementations.learnableDefinitionsRepository.PairsLearnableDefinitionsRepository
+import com.designdrivendevelopment.kotelok.repositoryImplementations.learnableDefinitionsRepository.WriterLearnableDefinitionsRepository
+import com.designdrivendevelopment.kotelok.repositoryImplementations.statisticsRepository.StatisticsRepositoryImpl
 import com.designdrivendevelopment.kotelok.screens.bottomNavigation.BottomNavigator
 import com.designdrivendevelopment.kotelok.screens.bottomNavigation.Tab
 import com.designdrivendevelopment.kotelok.screens.dictionaries.dictionariesScreen.DictionariesFragment
