+++ app/src/main/kotlin/com/designdrivendevelopment/kotelok/repositoryImplementations/editWordDefnititionsRepository/DefinitionsRequestResult.kt
@@ -1,4 +1,4 @@
-package com.designdrivendevelopment.kotelok
+package com.designdrivendevelopment.kotelok.repositoryImplementations.editWordDefnititionsRepository
 
 import com.designdrivendevelopment.kotelok.entities.WordDefinition
 
@@ -6,8 +6,8 @@ sealed class DefinitionsRequestResult {
     class Success(val definitions: List<WordDefinition>) : DefinitionsRequestResult()
 
     sealed class Failure(val errorMessage: String) : DefinitionsRequestResult() {
-        class HttpError(val code: Int, message: String) : DefinitionsRequestResult.Failure(message)
+        class HttpError(val code: Int, message: String) : Failure(message)
 
-        class Error(message: String) : DefinitionsRequestResult.Failure(message)
+        class Error(message: String) : Failure(message)
     }
 }
